consolidate dot scripts into dot.sh and refactor alias/function defs

+++ dot_backup.sh
@@ -3,5 +3,5 @@
 LOCATIONS="aliases bash_profile editorconfig exports functions gemrc gitconfig gitignore_global hushlogin psqlrc vimrc zshrc kerlrc history_preexec.sh p9k.zsh p10k.zsh"
 
 for location in $LOCATIONS; do
-	cp ~/.$location .$location
+  [ -f ~/.$location ] && cp ~/.$location .$location
 done
